8th done

+++ Unit2/8.c
@@ -2,7 +2,7 @@
 
 int main ()
 {
-	printf ("%u, %u: %u", ~5, (~0 >> 1), ~0 - (~0 >> 1));	
+	printf ("%u, %u: %u", ~5, (~0 >> 1), (unsigned int) ~0 - (unsigned int)(~0 >> 1));	
 	return 0;
 }
